+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ v0.4.0
   - Updated docstrings in numerous functions to ensure they render properly.
   - Add ``result["phase"] = (result["phase"] + 180) % 360 - 180`` to ``report.py`` to wrap ~360 degrees to ~0
   - Fixed a bug in ``readmat.py``
+  - Fix a bug that the internal initializer ``initialize_with_lm`` always uses the input method to initialize. Now it uses ``leastqs`` as the internal initializer.
 
 
 v0.3.9